requirements.txt
Pillow==10.3.0
# 可选：如需自动翻译，请单独安装 googletrans==4.0.0-rc1
